Move execSQLiteQuery function into utils
So that it can be re-used
