+++ lib/certificate.js
@@ -2,7 +2,7 @@ import crypto from 'crypto';
 import B from 'bluebird';
 import path from 'path';
 import { fs, mkdirp } from 'appium-support';
-import { exec } from 'teen_process';
+import { execSQLiteQuery } from './utils';
 
 const openssl = B.promisify(require('openssl-wrapper').exec);
 
@@ -130,9 +130,9 @@ class TrustStore {
     this.db = path.resolve(keychainsPath, 'TrustStore.sqlite3');
 
     // If it doesn't have a tsettings table, create one
-    await this.execSQLiteQuery(`CREATE TABLE IF NOT EXISTS tsettings (sha1 BLOB NOT NULL DEFAULT '', subj BLOB NOT NULL DEFAULT '', tset BLOB, data BLOB, PRIMARY KEY(sha1));`); 
+    await execSQLiteQuery(this.db, `CREATE TABLE IF NOT EXISTS tsettings (sha1 BLOB NOT NULL DEFAULT '', subj BLOB NOT NULL DEFAULT '', tset BLOB, data BLOB, PRIMARY KEY(sha1));`); 
     try {
-      await this.execSQLiteQuery('CREATE INDEX isubj ON tsettings(subj);');
+      await execSQLiteQuery(this.db, 'CREATE INDEX isubj ON tsettings(subj);');
     } catch (e) { }
     
 
@@ -144,9 +144,9 @@ class TrustStore {
    */
   async addRecord (sha1, tset, subj, data) {
     if (await this.hasRecords(subj)) {
-      return await this.execSQLiteQuery(`UPDATE tsettings SET sha1=x'?', tset='?', data=x'?' WHERE subj='?'`, sha1, tset, data, subj);
+      return await execSQLiteQuery(await this.getDB(), `UPDATE tsettings SET sha1=x'?', tset='?', data=x'?' WHERE subj='?'`, sha1, tset, data, subj);
     } else {
-      return await this.execSQLiteQuery(`INSERT INTO tsettings (sha1, subj, tset, data) VALUES (x'?', '?', '?', x'?')`, sha1, subj, tset, data);
+      return await execSQLiteQuery(await this.getDB(), `INSERT INTO tsettings (sha1, subj, tset, data) VALUES (x'?', '?', '?', x'?')`, sha1, subj, tset, data);
     }
   }
 
@@ -154,7 +154,7 @@ class TrustStore {
    * Remove record from tsettings
    */
   async removeRecord (subj) {
-    return await this.execSQLiteQuery(`DELETE FROM tsettings WHERE subj = '?'`, subj);
+    return await execSQLiteQuery(await this.getDB(), `DELETE FROM tsettings WHERE subj = '?'`, subj);
   }
 
   /**
@@ -165,28 +165,8 @@ class TrustStore {
   } 
 
   async getRecordCount (subj) {
-    let result =  (await this.execSQLiteQuery(`SELECT count(*) FROM tsettings WHERE subj = '?'`, subj)).stdout;
+    let result =  (await execSQLiteQuery(await this.getDB(), `SELECT count(*) FROM tsettings WHERE subj = '?'`, subj)).stdout;
     return parseInt(result.split('=')[1], 10);
-  }
-
-  escapeQuery (query) {
-    return query.replace(/'/g, "''");
-  }
-
-  /**
-   * Runs a command line sqlite3 query
-   */
-  async execSQLiteQuery (query, ...queryParams) {
-    let queryTokens = query.split('?');
-    let formattedQuery = [];
-    queryParams.forEach((param, i) => {
-      formattedQuery.push(queryTokens[i]);
-      formattedQuery.push(this.escapeQuery(param));
-    }); 
-    formattedQuery.push(queryTokens[queryTokens.length - 1]); 
-
-    let db = await this.getDB();
-    return await exec('sqlite3', ['-line', db, formattedQuery.join('')]);
   }
 }
