+++ lib/utils.js
@@ -205,4 +205,19 @@ async function uninstallSSLCert (pemText, udid) {
   }
 }
 
-export { killAllSimulators, endAllSimulatorDaemons, safeRimRaf, simExists, installSSLCert, uninstallSSLCert };
+/**
+ * Runs a command line sqlite3 query
+ */
+async function execSQLiteQuery (db, query, ...queryParams) {
+  let queryTokens = query.split('?');
+  let formattedQuery = [];
+  queryParams.forEach((param, i) => {
+    formattedQuery.push(queryTokens[i]);
+    formattedQuery.push(param.replace(/'/g, "''"));
+  }); 
+  formattedQuery.push(queryTokens[queryTokens.length - 1]); 
+
+  return await exec('sqlite3', ['-line', db, formattedQuery.join('')]);
+}
+
+export { killAllSimulators, endAllSimulatorDaemons, safeRimRaf, simExists, installSSLCert, uninstallSSLCert, execSQLiteQuery, enableCalendarAccess };
